changed to size_t and notebook added

+++ test.c
@@ -3,13 +3,13 @@
 
 int main()
 {
-    int n = 7;
+    size_t n = 7;
     double g[] = {.3, .2, .1, .0, -.1, -.2, -.3};
     double tolgrad = 0.15;
     
     truncate_gradient(g, n, tolgrad);
     
-    for(int i=0; i<n; i++){
+    for(size_t i=0; i<n; i++){
         printf("%4.1f\n", g[i]);
     }
 }
