+++ truncate_gradient.c
@@ -1,8 +1,8 @@
 #include "truncate_gradient.h"
 
-void truncate_gradient(double* g, int n, double tolgrad)
+void truncate_gradient(double* g, size_t n, double tolgrad)
 {
-    for(int i=0; i<n; i++){
+    for(size_t i=0; i<n; i++){
         if(fabs(g[i]) < tolgrad){
             g[i] = 0.0;
         }
